Add a Boolean to search for annotation replies
We're changing the behaviour of our NIPSA system (and the moderation
system we're building alongside it) so that annotations that have been
hidden will still be shown in a redacted form if someone has replied to
them.

Because this is adding a new field, rather than changing an existing
one, I think we should be able to update the settings in-place without
reindexing everything.